# api/API.yaml
swagger: '2.0'
info:
  title: AirControl API
  description: Control your Quadcopter with AirControl API
  version: "1.0.0"
# the domain of the service
host: api.copterapi.com
# array of all schemes that your API supports
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
  /user:
    get:
      summary: Get user's info
      description: ""
      parameters:
        - name: sessionID
          in: query
          description: ID of current session
          required: true
          type: string
      tags:
        - Users
      responses:
        200:
          description: User's info
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - Users
      summary: Add new User
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: User object that needs to be added to the Data Base
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Session'
        "405":
          description: Invalid input
    delete:
      tags:
        - Users
      summary: Add new User
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: User object that needs to deleted from the Data Base
          required: true
          schema:
            $ref: "#/definitions/DeleteUser"
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
      
definitions:

  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
        
  User:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
      email:
        type: string
        
  Session:
    type: object
    properties:
      sessionID:
        type: string
        
  DeleteUser:
    type: object
    properties:
      password:
        type: string
      email:
        type: string
      sessionID:
        type: string